requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Streamlit UI
streamlit==1.31.0

# Document Processing
unstructured==0.12.0
pymupdf==1.23.8
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Vector DB & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# LLM Integration
groq==0.4.2
ollama==0.1.6

# Utilities
pydantic==2.6.0
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1
httpx==0.26.0